requirements.txt
requests
beautifulsoup4
html2text
googletrans==4.0.0rc1
pyyaml
Pillow>=10.0.0
py2app>=0.28.0
